requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
requests
beautifulsoup4
sqlite-utils
openai 